Make DBus bus name settable on the command-line
The intention is that there will be multiple pulsevideo instances running
at any one time, enumerable by listing the current (and activateable)
bus-names, all with the prefix `com.stbtester.VideoSource.`.  With this
change the suffix can be set with the `--name` option.

We default to `dev_video0` as `/dev/video0` is the default device that
`v4l2src` uses, our default source-pipeline.  This also updates the
defaults of `dbusvideosourcesrc` as previously the default bus-name set
there was not the same as the hard-coded bus name in pulsevideo, leading
to the surprising result that it would not "just work" by default.