Fix potential printf issue: Hard code the format string
Previously the printf format string was being taken from `e.message`.  If
that string were to contain any `%` format specifiers (like `%i`) printf
would try to read the value from varargs, when there is none leading to
undefined behaviour.  This fixes that issue and makes the code a little
less obtuse in the process