Add GstTmpFileAllocator
Helper allocator for implementing zero-copy video IPC.  Uses the existing
GstDmabufAllocator to provide `GstMemory` that is backed by data stored
in a temporary file on `/dev/shm`.  The file-description of these files
can then be passed over a Unix domain socket to transfer the memory
between processes without copying it.