Add element fddepay
This element is the inverse counterpart to fdpay and the final piece to
enable zero-copy data transfer with GStreamer.

Note: fddepay is not yet secure.  For that to be the case it would need
to check the nature of the received FDs and ensure that they are either
sealed memfds or dmabuf backed before mapping them.

Conflicts:
	sys/tmpfile/Makefile.am