fdpay: Make it work with Gstreamer 1.2 on Ubuntu 14.04
... by implementing `transform_size` and `transform` rather than
`transform_ip`.  Usually `transform_ip` is used in cases where the output
buffers are the same size as the input buffers.  This isn't the case for
us, so implementing `transform` instead conforms to `GstBaseTransform`'s
expectations.

There was a change to how `GstBufferPool`s were used in the 1.3 series
that made the unusual way we were using `GstBaseTransform` work, but it
didn't work on GStreamer 1.2.  Now `fdpay` does.