tmpfileallocator: Use a different temporary filename for each frame
I suspect that pulsevideo is reusing the same inode over and over again
for each frame that it sends over the socket.  This commit will ensure
that a different filename is used each time we create a temporary file.
This filename will include the pid of the originating process and a
count of frames produced.  This should make it easier to debug and
determine whether the same file-descriptors are being re-used.