fdpay: Fix backing memory reuse
This fixes an issue where only one temporary file was created and reused
by pulsevideo.  When the FD was sent across the socket GStreamer would
return the `GstBuffer` to a `GstBufferPool` and it would then be reused
overwriting the data in the backing memory, even though it was still in
use by the downstream process.

This had the effect that even after the client had received a video frame
the contents would change.

Fixes github issues #1 and #3.

This is actually a reversion of a commit 589768b which made fdpay work
with GStreamer 1.2 (and thus Ubuntu 14.04). So this commit **breaks
pulsevideo on GStreamer 1.2**. This isn't an issue for my use-case as
where I use it the clients may be using GStreamer 1.2, but the pulsevideo
daemon runs with GStreamer 1.4.

Surprisingly this commit causes a small improvement to performance. I ran
the performance test 46 times both before and after and the results:

                     user     sys   total
Without this patch   2.40    0.17     2.6
With this patch      1.91    0.35     2.4

Each of those values is the median result for streaming 10s/250 frames of
720p25 video.

As expected the amount of time in the kernel has increased with this
commit, but the amount of time in userspace has decreased!

This reverts commit 589768be40e0d5fc88425ca659f85592292321d7.