tmpfileallocator: Use open(..., O_TMPFILE) if available
`O_TMPFILE` creates a file on a filesystem without giving it a name.  This
means that we don't `unlink`.  This is cleaner and potentially faster, as
there are fewer syscalls and we don't need to go through the `mkostemp`
find an unused filename dance.

`O_TMPFILE` was introduced in linux 3.11 (2013-09-02) and is not supported
on all filesystems so we include a fallback option here.  This is a kind
of half-way-house between normal files and memfds, although with none of
the security benefits of memfds.

This commit un-does the work introduced in b59a36a, but since we fixed
the backing memory reuse in the 7ac5d5e it's no longer necessary.