pulsevideosrc: Drop incorrectly sized raw video buffers
Sometimes I've seen v4l2src produce buffers that are smaller than you
would expect based on the caps.  I don't think it's technically an error,
but it can certainly surprise downstream elements.

I think it's not an error, but I'm not sure. This is more expedient than
going into detail and working out the "proper" answers to these questions.
Like: should v4l2src be producing buffers with invalid sizes? and does v4l
mark these buffers as broken?  Either way it's valid for pulsevideosrc to
do this as it's "supposed" to be workable over a security boundary so
doing this kind of validation is reasonable.

The reason I suspect that it's valid is that videoconvert doesn't seem to
complain too much when it's given a buffer like this.