tests: Make test_that_zerocopy_doesnt_leak_fds pass
Running `symmetry_test_assert_passthrough` twice before doing the initial
count of fds was sufficient to fix this test.  I've added
enable-last-sample=false to deallocate the fds sooner to clear up the
list of fds for easier debugging.

I don't know why I need to run the function twice, but it seems that the
process keeps a hold of the latest fd and the one before that.  If you only
send a single message through only one of the "slots" is filled so we see
an apparent leak.  The process also keeps a hold of the first fd we sent
through, but I don't know why or who the owner of that fd is.

I was debugging by inserting `system("ls -la /proc/$PPID/fd")` in various
places.