Protocol: Add capture_timestamp to fdpay header
The capture_timestamp is the time at which the video-frame was captured
(e.g. from v4l) as measured against the Linux CLOCK_MONOTONIC clock.  This
is useful for synchronisation between multiple sources.  the monotonic
clock was chosen because that is the format most v4l and alsa devices
provide their data in.

This is a backward incompatible change to the wire-protocol so I have
bumped the DBus interface version number.  Old versions and new versions
of pulsevideo will not interoperate, but at least they will fail early and
loudly.