pulsevideosrc: Ensure that buffers in tmpfile memory are read-only
This prevents clients from intefering with each other.  It's not secure
in that the FD sent across the socket is opened RW, so could be written
to but this should prevent accidental misuse.

You can still map the buffers with
`gst_buffer_map (..., GST_MAP_READWRITE)` because `GstBuffer` will
silently perform a copy on demand so clients shouldn't see any difference.