Report real caps rather than caps filter to clients
This will allow passing through information like h264 codec_data
over pulsevideo.  This information is required to understand the
h264 packets, but can't be predicted in advance.

We now only register the dbus name after we're ready to send data.  This
ensures we have caps as soon as an attach call comes in.  It's also the
right thing to do, we don't want to offer our service if we can't provide
one.