Revamp caps handling
Here we make things match the GStreamer model a bit better.  Our payloader
payloads the caps too so before it's:

    video/x-raw,format=BGR

and after it's

    application/x-fd,payloaded-name=video/x-raw,format=BGR

and vice-versa for the depayloader.  We're now communicating the payloaded
caps over DBus rather than the payloaded ones.

This way we can set the caps on the socketsrc, rather than having them
applied after the payloader using a capsfilter.  They will travel
downstream synchronised with the data.

This fits better with the GStreamer model and will make implementing
fddepay as a `GstBaseParse` possible.  Without this we don't know how to
set caps on the fddepay srcpad and GstBaseParse throws an error.