Add fault injection system
This allows us to inject faults into pulsevideo to deterministically create
failures at various points in the system.  It's controlled by environment
variables.  So now if I set:

    fdpay_buffer="skip skip usleep=1000000 skip abort" pulsevideo ...

The video stream will pause for 1s on the 3rd buffer and abort on the 4th
one.

The commands are:

* `usleep=<usecs>` - sleep for `<usecs>` microseconds.  Note: unlike with
  the other commands, after the sleep the next command will be executed.
* `skip` - don't inject a fault on this invocation
* `gerror` - return an error
* `abort` - crash with SIGABRT

The injection points are:

* `pre_attach` - called in the handler for the incoming DBus request before
  returning the value to the client.  This happens after the socket has
  been created and been added to the `multisocketsink`.
* `fdpay_buffer` - called in the streaming thread for every buffer to be
  sent.