DBus: Return error from Attach if caps not available
Without this if we don't have caps at `Attach` time the pulsevideo server
won't send any response to the `Attach` call, leading to the client hanging
for 25s until the DBus timeout is reached.  This was causing errors.
Instead now an error will be returned and the client will attempt to
reconnect.  When the reconnection is complete maybe it'll be fixed.

This is obviously still suboptimial, but it's a lot better than it was.