GstFdMemory: Support adjusting protection flags on memory as necessary
This is particularly important when using the
`GST_FD_MEMORY_FLAG_KEEP_MAPPED` flag.  I hit this issue in a pulsevideo
test after adding the `KEEP_MAPPED` flag.  The first time the memory was
mapped was in write only mode.  Then, when a downstream element came to
read it it would fail because it wasn't mapped with `PROT_READ`.