Don't use mmap below a certain size threshold
`mmap` is slower than copying for small buffers.  With this commit, if the
size of the buffer is below some threshold we'll use `write` and `read`
to payload/depayload.  We still use an FD transport in this case though.

I've set the threshold to 100KB, but it is customiseable via the
`TMPFILE_MMAP_THRESHOLD` environment variable so we can experiment with
different values easily.