pulsevideosink: Block `Attach` method until caps arrive
It turns out that our assumption that caps will be set by the time `Attach`
is called is false.  Now we no longer rely on that, instead we wait until
caps actually arrive.

I'm not 100% sure what thread this will be run from, (i.e. which thread
we'll be blocking with our sleeps) but I think it will be the GLib Mainloop
which should be ok as we'll only be sleeping a short amount of time in
practice.  I don't want to complicate this function any further with async,
etc.